Start refactoring plugin interface, issue #472
