+++ lib/impress.client.js
@@ -675,18 +675,18 @@ Client.prototype.fileHandler = function(handler, inheritance, callback) {
     if (exists) {
       client.runScript(handler, filePath, callback);
       if (!inheritance) application.cache.files[relPath] = impress.FILE_EXISTS;
-      application.watchCache(api.path.dirname(relPath));
+       application.cache.watch(api.path.dirname(relPath));
     } else {
       // Try to process request on parent directory
       if (client.execPath !== '/' && handler !== 'meta') {
         client.execPath = api.impress.dirname(client.execPath);
         client.execPathDir = application.dir + '/server' + client.execPath;
         client.fileHandler(handler, inheritance, callback);
-        application.watchCache('/server' + client.execPath);
+         application.cache.watch('/server' + client.execPath);
       } else {
         // Lose hope to execute request and drop connection
         if (!inheritance) application.cache.files[relPath] = impress.FILE_NOT_FOUND;
-        application.watchCache(api.path.dirname(relPath));
+         application.cache.watch(api.path.dirname(relPath));
         callback();
       }
     }
@@ -715,7 +715,7 @@ Client.prototype.detectRealPath = function(callback) {
       relPath = application.relative(client.realPathDir);
       client.detectRealPath(callback);
     }
-    application.watchCache(api.impress.stripTrailingSlash(relPath));
+     application.cache.watch(api.impress.stripTrailingSlash(relPath));
   });
 };
 
@@ -867,7 +867,7 @@ Client.prototype.serveStatic = function(relPath, onNotServed) {
   function serveFile(err, stats) {
     if (err) {
       //application.cache.static[relPath] = impress.FILE_NOT_FOUND;
-      application.watchCache(api.path.dirname(relPath));
+       application.cache.watch(api.path.dirname(relPath));
       var isIndex = api.impress.endsWith(relPath, '/index.html');
       if (client.path !== '/' && isIndex) {
         client.index(api.path.dirname(filePath));
