+++ lib/impress.cloud.js
@@ -239,7 +239,7 @@ impress.cloud.generateNextCluster = function() {
 
 // Mixin cloud to application instance
 //
-impress.cloud.mixinApplication = function(application) {
+impress.cloud.mixin = function(application) {
 
   // Call remote method
   //
