+++ lib/impress.constants.js
@@ -24,7 +24,13 @@ impress.FILE_PARSE_ERROR = 13;
 impress.DIR_EXISTS = 50;
 impress.DIR_NOT_EXISTS = 51;
 
-impress.CORE_PLUGINS = ['security', 'state', 'cloud', 'sse', 'websocket', 'rpc'];
+impress.CORE_PLUGINS = [
+  'log', 'cache',
+  'application', 'client', 'index', 'files', 'preprocess',
+  'security', 'state', 'cloud', 'cloud.tcp',
+  'sse', 'websocket', 'rpc', 'health', 'waf'
+];
+
 impress.HTTP_VEBS = ['get', 'post', 'put', 'delete', 'patch', 'head', 'options'];
 
 impress.UPLOAD_SIZE_ZIP = 1048576;
