+++ lib/impress.health.js
@@ -1,8 +1,6 @@
 'use strict';
 
 // Health monitoring for Impress Application Server
-//
-impress.health = {};
 
 // Init health monitoring
 //
