+++ lib/impress.js
@@ -7,19 +7,7 @@ global.impress = new api.events.EventEmitter();
 
 impress.isMainApplication = true;
 
-// Impress core modules
-//
 require('./impress.constants');
-require('./impress.application');
-require('./impress.client');
-require('./impress.templating');
-require('./impress.index');
-require('./impress.files');
-require('./impress.security');
-require('./impress.health');
-require('./impress.cloud');
-require('./impress.preprocess');
-require('./impress.waf');
 require('./db');
 
 // Paths to directories
@@ -85,23 +73,31 @@ impress.nextWorkerId = 1;
 // Load plugins
 //
 impress.loadPlugins = function() {
-  var pluginName, pluginPath, cache;
+  var plugin, pluginName, pluginPath;
+  console.log('to: impress');
   for (var i = 0, len = impress.CORE_PLUGINS.length; i < len; i++) {
     pluginName = impress.CORE_PLUGINS[i];
     pluginPath = './impress.' + pluginName + '.js';
-    cache = require.cache[require.resolve(pluginPath)];
-    if (!cache) require(pluginPath);
+    impress[pluginName] = {};
+    require(pluginPath);
+    plugin = impress[pluginName];
+    console.log('  load: '+pluginName+' plugin: '+typeof(plugin)+' mixin: '+(plugin ? typeof(plugin.mixin) : '-'));
+    if (pluginName ==='log' || pluginName ==='application' || pluginName ==='cache') plugin.mixin(impress);
+    //if (plugin && plugin.mixin) plugin.mixin(impress);
   }
 };
 
 // Mixin plugins
 //
 impress.mixinPlugins = function(application) {
+  console.log('application: '+application.name);
   var plugin, pluginName;
   for (var i = 0, len = impress.CORE_PLUGINS.length; i < len; i++) {
     pluginName = impress.CORE_PLUGINS[i];
     plugin = impress[pluginName];
-    plugin.mixinApplication(application);
+    application[pluginName] = {};
+    console.log('  mix: '+pluginName+' plugin: '+typeof(plugin)+' mixin: '+(plugin ? typeof(plugin.mixin) : '-'));
+    if (plugin && plugin.mixin) plugin.mixin(application);
   }
 };
 
@@ -211,21 +207,18 @@ impress.loadApplication = function(appName, dir, callback) {
   var application = new api.events.EventEmitter();
   application.name = appName;
   application.dir = dir;
-  impress.mixinApplication(application);
+  impress.mixinPlugins(application);
   impress.applications[application.name] = application;
   application.createSandbox(function() {
     application.loadConfig(function() {
       application.preprocessConfig();
-      impress.log.mixinApplication(application);
+      application.log.init();
       application.log.open(function() {
         application.emit('start');
         api.db.openApplicationDatabases(application, function() {
-          impress.mixinPlugins(application);
           application.loadPlaces(function() {
             application.emit('started');
-            if (application.config.application && application.config.application.preload) {
-              application.preloadDirectory('/');
-            }
+            // application.cache.preload('/');
             callback();
           });
         });
@@ -267,19 +260,16 @@ var isFirstStart = true;
 // Start servers
 //
 impress.server.start = function() {
-  impress.mixinApplication(impress);
+  impress.loadPlugins();
+  impress.mode = process.env.IMPRESS_MODE || '';
+  impress.workerId = api.cluster.isMaster ? 0 : process.env.WORKER_ID;
+  impress.workerType = process.env.WORKER_TYPE;
+  impress.serverName = process.env.WORKER_SERVER_NAME;
+  if (api.cluster.isMaster) console.log('Impress Application Server'.green.bold + ' starting, reading configuration'.green);
   impress.createSandbox(function() {
-    impress.mode = process.env.IMPRESS_MODE || '';
-    impress.workerId = api.cluster.isMaster ? 0 : process.env.WORKER_ID;
-    impress.workerType = process.env.WORKER_TYPE;
-    impress.serverName = process.env.WORKER_SERVER_NAME;
-    if (api.cluster.isMaster) console.log('Impress Application Server'.green.bold + ' starting, reading configuration'.green);
     impress.loadConfig(function() {
       impress.preprocessConfig();
-      require('./impress.log');
-      require('./impress.cloud.' + impress.config.scale.transport);
-      impress.loadPlugins();
-      impress.log.mixinApplication(impress);
+      impress.log.init();
       impress.log.open(function() {
         if (impress.workerType === 'long') {
           impress.nodeId = impress.config.scale.cluster + 'L' + impress.workerId;
