+++ lib/impress.passport.js
@@ -10,16 +10,13 @@
 //   passport-facebook
 //   passport-twitter
 
-impress.passport = {};
-
 // Passport plugin definition
 //
-impress.passport.mixinApplication = function(application) {
+impress.passport.mixin = function(application) {
 
   var providers = api.impress.getByPath(application, 'config.passport.strategies');
   if (!providers || !api.passport) return;
 
-  application.passport = {};
   var provider, providerName, providerId, strategies = {},
       handlers = [api.passport.initialize(), api.passport.session()];
   for (providerName in providers) {
