+++ lib/impress.preprocess.js
@@ -1,8 +1,6 @@
 'use strict';
 
 // SASS and JsUglify preprocessors for Impress Application Server
-//
-impress.preprocess = {};
 
 api.sass = api.impress.require('node-sass');
 api.uglify = api.impress.require('uglify-js');
