+++ lib/impress.security.js
@@ -1,8 +1,6 @@
 'use strict';
 
 // Security plugin for Impress Application Server
-//
-impress.security = {};
 
 // User class
 //   user - an object from storage provider we need to copy all parameters from
@@ -45,7 +43,10 @@ impress.security.getDatabaseProvider = function(driverName) {
 
 // Mixin security to application instance
 //
-impress.security.mixinApplication = function(application) {
+impress.security.mixin = function(application) {
+
+  application.User = impress.security.User;
+  application.Session = impress.security.Session;
 
   var singleWarning = true;
 
@@ -57,10 +58,6 @@ impress.security.mixinApplication = function(application) {
       singleWarning = false;
     }
   }
-
-  application.security = {};
-  application.User = impress.security.User;
-  application.Session = impress.security.Session;
 
   // Get user for given session or {}
   //
@@ -150,7 +147,7 @@ impress.security.mixinApplication = function(application) {
   //
   if (application.databases && application.databases.security) {
     var provider = impress.security.getDatabaseProvider(application.databases.security.schema);
-    if (provider) provider.mixinApplication(application);
+    if (provider) provider.mixin(application);
   }
 
 };
