+++ lib/impress.security.mongodb.js
@@ -6,7 +6,7 @@ impress.security.mongodb = {};
 
 // Mixin security to application instance
 //
-impress.security.mongodb.mixinApplication = function(application) {
+impress.security.mongodb.mixin = function(application) {
 
   application.security.hasDbWarned = false;
