+++ lib/impress.sse.js
@@ -1,25 +1,21 @@
 'use strict';
 
 // Server-Sent Events plugin for Impress Application Server
-//
-impress.sse = {};
 
 var SSE_CONNECTED = new Buffer(':' + new Array(2049).join(' ') + '\n\n', 'utf8'),
     SSE_SET_RETRY = new Buffer('retry: 2000\n\n', 'utf8'),
     SSE_HEARTBEAT = new Buffer(':\n\n', 'utf8'),
     SSE_HB_INTERVAL = 10000;
 
-impress.sse.mixinApplication = function(application) {
+impress.sse.mixin = function(application) {
 
-  application.sse = {
-    nextConnectionId: 1, // counter to be used as connection identifier
-    connections: {},     // all SSE connections
-    statistics: {
-      incoming: 0,     // incoming connection count from server start
-      active: 0,       // active connection count
-      disconnected: 0, // disconnected connection count from server start
-      errors: 0        // connection error count from server start
-    }
+  application.sse.nextConnectionId = 1, // counter to be used as connection identifier
+  application.sse.connections = {},     // all SSE connections
+  application.sse.statistics = {
+    incoming: 0,     // incoming connection count from server start
+    active: 0,       // active connection count
+    disconnected: 0, // disconnected connection count from server start
+    errors: 0        // connection error count from server start
   };
 
   application.sse.heartbeat = setInterval(function() {
