+++ lib/impress.state.js
@@ -1,14 +1,15 @@
 'use strict';
 
 // Application state synchronization plugin for Impress Application Server
-//
-impress.state = {};
 
 // Stateful Application plugin definition
 //
-impress.state.mixinApplication = function(application) {
+impress.state.mixin = function(application) {
 
-  var state = new api.events.EventEmitter();
+  application.state = new api.events.EventEmitter();
+
+  var state = application.state;
+
   state.data = {};         // data structures for synchronization
   state.delta = {};        // delta to be synced at next syncronization
   state.back = {};         // data for rollback
@@ -59,10 +60,6 @@ impress.state.mixinApplication = function(application) {
       }
     }
   }
-
-  // Access state from application
-  //
-  application.state = state;
 
   // Start state transaction
   //
