+++ lib/impress.templating.js
@@ -41,18 +41,18 @@ Client.prototype.template = function(data, file, cursor, callback) {
     } else if (result) {
       client.include(data, filePath, cursor, callback);
       application.cache.files[relPath] = impress.FILE_EXISTS;
-      application.watchCache(api.path.dirname(relPath));
+       application.cache.watch(api.path.dirname(relPath));
     } else {
       // Try to find template in parent directory
       if (client.execPath !== '/' && client.execPath !== '/server' && client.execPath !== '.') {
         client.execPath = api.impress.dirname(client.execPath);
         client.execPathDir = application.dir + '/server' + client.execPath;
         client.template(data, file, cursor, callback);
-        application.watchCache('/server' + api.impress.stripTrailingSlash(client.execPath));
+         application.cache.watch('/server' + api.impress.stripTrailingSlash(client.execPath));
       } else {
         // Lose hope to find template and save cache
         application.cache.files[relPath] = impress.FILE_NOT_FOUND;
-        application.watchCache(relPath);
+         application.cache.watch(relPath);
         callback(impress.TPL_NOT_FOUND + relPath);
       }
     }
@@ -88,7 +88,7 @@ Client.prototype.include = function(data, filePath, cursor, callback) {
       }
     });
     filePath = application.relative(api.path.dirname(filePath));
-    application.watchCache(filePath);
+     application.cache.watch(filePath);
   }
 };
