+++ lib/impress.waf.js
@@ -1,8 +1,6 @@
 'use strict';
 
 // Web Application Firewall for Impress Application Server
-//
-impress.waf = {};
 
 impress.waf.ACCESS_ALLOWED = 0;
 impress.waf.ACCESS_DENIED = 1;
@@ -176,14 +174,15 @@ impress.waf.mixins.limit = function(item) {
 };
 
 // Mixin WAF functionality to WAF objects
+//   TODO: hints for: items, mixins
 //
-impress.waf.mixin = function(items, mixins) {
+impress.waf.mx = function(items, mixins) {
   mixins.map(function(mixin) {
     items.map(mixin);
   });
 };
 
-impress.waf.mixin([
+impress.waf.mx([
   impress.waf.ip,
   impress.waf.sid,
   impress.waf.host
@@ -192,7 +191,7 @@ impress.waf.mixin([
   impress.waf.mixins.limit
 ]);
 
-impress.waf.mixin([
+impress.waf.mx([
   impress.waf.url,
   impress.waf.app,
   impress.waf.srv
