+++ lib/impress.websocket.js
@@ -4,67 +4,61 @@
 //
 api.websocket = api.impress.require('websocket');
 
-if (api.websocket) {
+impress.websocket = api.websocket;
 
-  impress.websocket = api.websocket;
+// Upgrade HTTP server to support websocket
+//
+impress.websocket.upgradeServer = function(server) {
 
-  // Upgrade HTTP server to support websocket
+  server.webSocketServer = new api.websocket.server({
+    httpServer: server,
+    autoAcceptConnections: false
+  });
+
+  server.webSocketServer.on('request', function(request) {
+    var req = request.httpRequest,
+        res = request.socket;
+    res.websocket = {};
+    res.websocket.request = request;
+
+    var client = impress.dispatcher(req, res),
+        application = client.application;
+    application.emit('connect', client);
+  });
+
+};
+
+// Websocket Application plugin definition
+//
+impress.websocket.mixin = function(application) {
+
+  // Websocket initialization (to be called automatically by Impress core)
   //
-  impress.websocket.upgradeServer = function(server) {
+  application.websocket.initialize = function(client) {
 
-    server.webSocketServer = new api.websocket.server({
-      httpServer: server,
-      autoAcceptConnections: false
-    });
+    if (client.res.websocket) client.websocket = client.res.websocket;
+    else {
+      client.websocket = { fake: true };
+      client.error(400);
+    }
 
-    server.webSocketServer.on('request', function(request) {
-      var req = request.httpRequest,
-          res = request.socket;
-      res.websocket = {};
-      res.websocket.request = request;
-
-      var client = impress.dispatcher(req, res),
-          application = client.application;
-      application.emit('connect', client);
-    });
-
-  };
-
-  // Websocket Application plugin definition
-  //
-  impress.websocket.mixinApplication = function(application) {
-
-    application.websocket = {};
-
-    // Websocket initialization (to be called automatically by Impress core)
+    // Accept websocket connection
     //
-    application.websocket.initialize = function(client) {
-
-      if (client.res.websocket) client.websocket = client.res.websocket;
-      else {
-        client.websocket = { fake: true };
-        client.error(400);
-      }
-
-      // Accept websocket connection
-      //
-      client.websocket.accept = function() {
-        client.websocket.isAccepted = true;
-        client.res.statusCode = 101;
-        client.accessLog();
-        if (client.websocket.fake) return;
-        return client.websocket.request.accept('', client.websocket.request.origin);
-      };
-
-    };
-
-    // Finalize websocket connection
-    //
-    application.websocket.finalize = function(client) {
-      var ws = client.websocket;
-      if (ws && !ws.isAccepted) ws.request.reject();
+    client.websocket.accept = function() {
+      client.websocket.isAccepted = true;
+      client.res.statusCode = 101;
+      client.accessLog();
+      if (client.websocket.fake) return;
+      return client.websocket.request.accept('', client.websocket.request.origin);
     };
 
   };
 
-}
+  // Finalize websocket connection
+  //
+  application.websocket.finalize = function(client) {
+    var ws = client.websocket;
+    if (ws && !ws.isAccepted) ws.request.reject();
+  };
+
+};
